refactor: move authentication domain types and GrantTypeService into ferriskey-domain
