+++ core/src/domain/authentication/ports.rs
@@ -10,8 +10,8 @@ use crate::domain::{
         entities::{
             AuthInput, AuthOutput, AuthSession, AuthenticateInput, AuthenticateOutput,
             AuthenticationError, AuthorizeRequestInput, AuthorizeRequestOutput,
-            CredentialsAuthParams, ExchangeTokenInput, GrantType, JwtToken,
-            TokenIntrospectionResponse, WebAuthnChallenge,
+            CredentialsAuthParams, ExchangeTokenInput, JwtToken, TokenIntrospectionResponse,
+            WebAuthnChallenge,
         },
         value_objects::{
             AuthenticationResult, CreateAuthSessionRequest, GrantTypeParams, RegisterUserInput,
@@ -22,18 +22,9 @@ use crate::domain::{
     jwt::entities::JwkKey,
 };
 
-/// A strategy for handling different OAuth2 grant types during authentication.
-///
-/// This trait defines the contract for implementing specific grant type strategies,
-/// such as `AuthorizationCode`, `ClientCredentials`, or `Password` grant types.
-/// Each implementation of this trait should handle the logic for its respective grant type.
-pub trait GrantTypeService: Send + Sync {
-    fn authenticate_with_grant_type(
-        &self,
-        grant_type: GrantType,
-        params: GrantTypeParams,
-    ) -> impl Future<Output = Result<JwtToken, AuthenticationError>> + Send;
-}
+// `GrantTypeService` now lives in `ferriskey-domain` (its signature touches only plain types).
+// Re-exported so existing `crate::domain::authentication::ports::GrantTypeService` sites resolve.
+pub use ferriskey_domain::authentication::ports::GrantTypeService;
 
 pub trait AuthSessionService: Send + Sync {
     fn create_session(
