+++ libs/ferriskey-domain/Cargo.toml
@@ -8,6 +8,7 @@ edition.workspace = true
 chrono = { version = "0.4.43", features = ["serde"] }
 rand = "0.8"
 serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
 utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
 uuid = { version = "1.20.0", features = ["serde", "v4", "v7"] }
 thiserror = "2.0.12"
