+++ libs/ferriskey-domain/src/authentication/mod.rs
@@ -1,1 +1,3 @@
 pub mod entities;
+pub mod ports;
+pub mod value_objects;
